Npm repo, JSDoc refinements
This is now an npm project. It does not have to be used as such, you can
just copy these files to the using project or include it as a git
submodule. This is just because I would like JSDoc explicitely installed
as a dependency and later also linting stuff, so anyone can quickly
generate the docs when needed.